backend/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
black==18.6b4
cffi==1.11.5
click==6.7
cryptography==2.3.1
Django==2.1.1
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.5
six==1.11.0
toml==0.9.6
